+++ packages/browser-extension/src/modules/ui/hooks/use-event-handlers.ts
