+++ packages/browser-extension/src/modules/ui/hooks/use-fully-visible.ts
