+++ packages/browser-extension/src/modules/ui/hooks/use-is-offline.ts
