+++ packages/browser-extension/src/modules/ui/utils/clipboard.ts
