+++ packages/browser-extension/src/modules/ui/utils/dom.ts
