+++ packages/browser-extension/src/modules/ui/utils/recursive-poller.ts
